feat: add file upload and create pdf on orders

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const Product = require("../models/product.js");
 
+const fileHelper = require("../util/file");
+
 const { validationResult } = require("express-validator");
 
 exports.getAddProduct = (req, res, next) => {
@@ -15,27 +17,45 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.file;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-  console.log(imageUrl);
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
+  console.log(image);
+  if (!image) {
     return res.status(422).render("admin/edit-product", {
-      pageTitle: "Edit Product",
+      pageTitle: "Add Product",
       path: "/admin/add-product",
       editing: false,
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
+        price: price,
+        description: description,
+      },
+      errorMessage: "Attached file is not an image.",
+      validationErrors: [],
+    });
+  }
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    console.log(errors.array());
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
         price: price,
         description: description,
       },
       errorMessage: errors.array()[0].msg,
-      validationsErrors: errors.array(),
+      validationErrors: errors.array(),
     });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product({
     title: title,
@@ -46,29 +66,14 @@ exports.postProduct = (req, res, next) => {
   });
   product
     .save()
-    .then(() => {
+    .then((result) => {
       console.log("Create the Product");
       res.redirect("/admin/products");
     })
     .catch((err) => {
-      /* return res.status(500).render("admin/edit-product", {
-        pageTitle: "Edit Product",
-        path: "/admin/add-product",
-        editing: false,
-        hasError: true,
-        product: {
-          title: title,
-          imageUrl: imageUrl,
-          price: price,
-          description: description,
-        },
-        errorMessage: "Database operation failed",
-        validationsErrors: [],
-      }); 
-      res.redirect("/500");*/
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(500);
+      return next(error);
     });
 };
 
@@ -96,14 +101,14 @@ exports.getEditProduct = (req, res, next) => {
     .catch((err) => {
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(500);
+      return next(error);
     });
 };
 
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedPrice = req.body.price;
   const updatedDescription = req.body.description;
 
@@ -116,13 +121,12 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDescription,
         _id: prodId,
       },
       errorMessage: errors.array()[0].msg,
-      validationsErrors: errors.array(),
+      validationErrors: errors.array(),
     });
   }
 
@@ -132,19 +136,21 @@ exports.postEditProduct = (req, res, next) => {
         return res.redirect("/");
       }
       product.title = updatedTitle;
-      product.imageUrl = updatedImageUrl;
       product.price = updatedPrice;
       product.description = updatedDescription;
-      return product.save().then(() => {
-        console.log("Updated Product");
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
+      return product.save().then((result) => {
+        console.log("UPDATED PRODUCT!");
         res.redirect("/admin/products");
       });
     })
-
     .catch((err) => {
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(500);
+      return next(error);
     });
 };
 
@@ -161,19 +167,26 @@ exports.getProducts = (req, res, next) => {
     .catch((err) => {
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(500);
+      return next(error);
     });
 };
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found"));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       res.redirect("/admin/products");
     })
     .catch((err) => {
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(500);
+      return next(error);
     });
 };
